add treue angle format data

+++ ShuttleManager/ShuttleManager.Shared/Models/Shuttle.cs
@@ -33,7 +33,7 @@ public class Shuttle
     public DateTime LastActivity { get; set; } = DateTime.Now;
 
     public double Temperature { get; set; } = 0.0;
-    public int Angle { get; set; } = 0;
+    public double Angle { get; set; } = 0;
     public int Length { get; set; } = 0;
     public int Position { get; set; } = 0;
     public string FifoLifoMode { get; set; } = "Unknown";
